refactor: 프로필 섹션 컴포넌트 분리 (#146)
