+++ src/components/profile/index.ts
@@ -1,0 +1,3 @@
+import ProfileSection from "@/components/profile/ProfileSection";
+
+export { ProfileSection };
